requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
matplotlib>=3.8.2
pydantic>=2.5.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
email-validator>=2.1.0
